% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastLmMatrix.R
\name{fastLmMatrix}
\alias{fastLmMatrix}
\alias{fastLmMatrix.default}
\alias{fastLmMatrix.formula}
\title{Fit Multivariate Linear Models}
\usage{
fastLmMatrix(x, ...)

\method{fastLmMatrix}{default}(x, y, tol = 1e-07, method = 0L, ...)

\method{fastLmMatrix}{formula}(formula, data = list(), method = 0L, ...)
}
\arguments{
\item{x}{a model matrix \eqn{X}.}

\item{...}{optional parameters passed to methods.}

\item{y}{the response matrix \eqn{Y}.}

\item{tol}{tolerance for the rank calculation.}

\item{method}{an integer with value \code{0} for the column-pivoted QR decomposition, \code{1} for the unpivoted QR decomposition, \code{2} for the LLT Cholesky, \code{3} for the LDLT Cholesky, and \code{4} for the Jacobi singular value decomposition (SVD). Default is zero.}

\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by \code{as.data.frame} to a data frame) containing the variables in the model.}
}
\value{
\code{fastLmMatrix()} returns a list with the following components:
\item{coefficients}{\eqn{p \times k} matrix of coefficients.}
\item{fitted.values}{\eqn{n \times k} matrix of fitted values.}
\item{residuals}{\eqn{n \times k} matrix of residuals.}
\item{rank}{an integer giving the numeric rank of the model matrix \eqn{X}.}
}
\description{
\code{fastLmMatrix()} estimates the linear model for multivariate response using one of several methods implemented using the \code{Eigen} linear algebra library.
}
\details{
\code{fastLmMatrix()} is a performance-optimized version of the standard \code{lm.fit()} function. Unlike \code{RcppEigen::fastLm()}, it is specifically designed to handle multivariate responses (\eqn{Y} as a matrix).
It leverages the Eigen C++ template library for high-performance linear algebra, providing several decomposition methods with different trade-offs between speed and numerical stability.
}
\seealso{
\code{\link[stats]{lm.fit}}, \code{\link[RcppEigen]{fastLm}}
}
