% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_surrogate.R
\name{global_surrogate}
\alias{global_surrogate}
\title{Global Surrogate Model Specification}
\usage{
global_surrogate(mode = "regression", penalty = NULL, num_knots = NULL)
}
\arguments{
\item{mode}{A single character string for the type of model.}

\item{penalty}{A non-negative number representing the total amount of regularization (`lambda` in midr).}

\item{k}{An integer for the maximum number of sample points (`k` in midr).}
}
\description{
\code{global_surrogate()} defines a model that serves as a surrogate model of another target model. This function can fit classification and regression models.
}
