% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_perspective.R
\name{mid_perspective}
\alias{mid_perspective}
\title{Perspective Plot of MID Effects}
\usage{
mid_perspective(
  object,
  xvar,
  yvar = NULL,
  x = NULL,
  y = NULL,
  ...,
  resolution = 50L
)
}
\arguments{
\item{object}{a "mid" object, typically the result of \code{midr::interpret()}.}

\item{xvar}{a character string with the name of the variable for the x-axis. Alternatively, a single string in the format \code{xvar:yvar} can be provided, in which case \code{yvar} can be omitted.}

\item{yvar}{a character string with the name of the variable for the y-axis.}

\item{x}{a numeric vector specifying the sequence of values for the x-axis.}

\item{y}{a numeric vector specifying the sequence of values for the y-axis.}

\item{...}{additional arguments passed on to \code{graphics::persp()}. Used to customize the plot's appearance, such as view angles (\code{theta}, \code{phi}) or color (\code{col}).}

\item{resolution}{an integer or an integer vector of length two, specifying the number of grid points to use when creating the sequence for numeric variables.}
}
\value{
\code{mid_perspective()} invisibly returns the viewing transformation matrix, see \code{\link[graphics]{persp}} for details.
This function is primarily called for its side effect of creating a plot.
}
\description{
Visualizes the combined effect of two variables from a "mid" object using a 3D perspective plot.
}
\details{
This function calculates the sum of the main effects of \code{xvar} and \code{yvar} and their interaction effect (\code{xvar:yvar}).
The resulting sum is plotted as the height on the z-axis.
}
\examples{
mid <- midr::interpret(mpg ~ wt * hp + am, data = mtcars, lambda = .5)

# Create a basic perspective plot
mid_perspective(mid, xvar = "wt", yvar = "hp")

# Customize the plot by passing arguments to graphics::persp
mid_perspective(mid, "wt", "hp", theta = 210, phi = 20, col = "lightblue", shade = .5)
}
\seealso{
\code{\link[graphics]{persp}}
}
