% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_reg.R
\name{mid_reg}
\alias{mid_reg}
\alias{fit_mid_reg}
\title{MID Regression}
\usage{
mid_reg(
  mode = "regression",
  engine = "midr",
  params_main = NULL,
  params_inter = NULL,
  penalty = NULL,
  custom_formula = NULL
)

fit_mid_reg(
  x,
  y,
  weights = NULL,
  params_main = NA,
  params_inter = NA,
  penalty = 0,
  custom_formula = NULL,
  ...
)
}
\arguments{
\item{mode}{A single character string for the type of model.}

\item{params_main}{An integer for the maximum number of sample points to model main effects. i.e., \code{k} for main effects in \code{interpret()}.}

\item{params_inter}{An integer for the maximum number of sample points to model interaction effects. i.e., \code{k} for interactions in \code{interpret()}.}

\item{penalty}{A non-negative number representing the total amount of regularization. i.e., \code{lambda} in \code{interpret()}.}

\item{custom_formula}{A formula object. If passed, \code{fit()} internally uses \code{interpret.formula()} with the specified formula, otherwise it uses \code{interpret.default()}. i.e., `formula` in \code{interpret.formula()}.}

\item{x}{a data frame or matrix of predictor variables.}

\item{y}{a vector of target variable.}

\item{weights}{an optional vector of case weights.}

\item{...}{other arguments passed to \code{midr::interpret()}.}
}
\value{
a \code{mid_reg} model specification.
}
\description{
\code{mid_reg()} defines a model that can predict numeric values from predictors using a set of functions each with up to two predictors.
This function can fit regression models.
}
\details{
This function is the main specification for the \strong{parsnip} model.
The underlying fitting is performed by \code{fit_mid_reg()}, which is a helper function that connects \code{fit()} to the \code{midr::interpret()} function.
}
\seealso{
\code{\link[midr]{interpret}}
}
\keyword{internal}
