% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_surrogate.R
\name{mid_surrogate}
\alias{mid_surrogate}
\alias{fit_mid_surrogate}
\title{Global Surrogate Model Specification}
\usage{
mid_surrogate(
  mode = "regression",
  params_main = NULL,
  params_inter = NULL,
  penalty = NULL,
  custom_formula = NULL
)

fit_mid_surrogate(
  x,
  y,
  weights = NULL,
  params_main = NA,
  params_inter = NA,
  penalty = 0,
  custom_formula = NULL,
  ...
)
}
\arguments{
\item{mode}{A single character string for the type of model.}

\item{params_main}{An integer for the maximum number of sample points to model main effects (\code{k} for main effects in \code{interpret()}).}

\item{params_inter}{An integer for the maximum number of sample points to model interaction effects (\code{k} for interactions in \code{interpret()}).}

\item{penalty}{A non-negative number representing the total amount of regularization (\code{lambda} in \code{interpret()}).}

\item{custom_formula}{A formula object. If passed, \code{fit()} internally uses \code{interpret.formula()}, otherwise it uses \code{interpret.default()} (`formula` in \code{interpret()}).}

\item{x}{a data frame or matrix of predictor variables.}

\item{y}{a vector of target variable.}

\item{weights}{an optional vector of case weights.}

\item{...}{other arguments passed to \code{midr::interpret()}.}
}
\value{
a \code{mid_surrogate} model specification.
}
\description{
\code{mid_surrogate()} defines a model that serves as a surrogate model of another target model. This function can fit classification (to be implemented) and regression models.
}
\details{
This function is the main specification for the \strong{parsnip} model.
The underlying fitting is performed by \code{fit_mid_surrogate()}, which is a helper function that connects \code{parsnip::fit()} to the \code{midr::interpret()} function.
}
\seealso{
\code{\link[midr]{interpret}}
}
\keyword{internal}
