% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midnight-package.R
\docType{package}
\name{midnight-package}
\alias{midnight}
\alias{midnight-package}
\title{midnight: Parsnip Engine for Maximum Interpretation Decomposition Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a 'parsnip' engine for the 'midr' package, enabling users to fit, tune, and evaluate Maximum Interpretation Decomposition (MID) models within the 'tidymodels' framework. Developed through research by the 'Moonlight Seminar 2025', a study group of actuaries from the Institute of Actuaries of Japan, to enhance practical applications of interpretable modeling. Detailed methodology is available in Asashiba et al. (2025) \doi{10.48550/arXiv.2506.08338}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ryo-asashi/midnight}
  \item \url{https://ryo-asashi.github.io/midnight/}
  \item Report bugs at \url{https://github.com/ryo-asashi/midnight/issues}
}

}
\author{
\strong{Maintainer}: Ryoichi Asashiba \email{ryoichi.asashiba@gmail.com}

}
\keyword{internal}
