% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persp_mid.R
\name{persp.mid}
\alias{persp.mid}
\title{Perspective Plot of MID Effects}
\usage{
\method{persp}{mid}(object, xvar, yvar = NULL, ..., xval = NULL, yval = NULL)
}
\arguments{
\item{object}{a "mid" object, typically the result of \code{midr::interpret()}.}

\item{xvar}{a character string with the name of the variable for the x-axis. Alternatively, a single string in the format \code{xvar:yvar} can be provided, in which case \code{yvar} can be omitted.}

\item{yvar}{a character string with the name of the variable for the y-axis.}

\item{...}{additional arguments passed on to \code{graphics:::persp.default()}. Used to customize the plot's appearance, such as view angles (\code{theta}, \code{phi}) or color (\code{col}).}

\item{xval}{a numeric or character vector specifying the sequence of values for the x-axis.}

\item{yval}{a numeric or character vector specifying the sequence of values for the y-axis.}
}
\value{
\code{persp.mid()} invisibly returns the viewing transformation matrix, see \code{\link[graphics]{persp}} for details.
This function is primarily called for its side effect of creating a plot.
}
\description{
Visualizes the combined effect of two variables from a "mid" object using a 3D perspective plot.
}
\details{
This is an S3 method for the \code{persp()} generic that calculates the sum of the main effects of \code{xvar} and \code{yvar} and their interaction effect (\code{xvar:yvar}).
The resulting sum is plotted as the height on the z-axis.
}
\examples{
mid <- midr::interpret(mpg ~ wt * hp + factor(am), data = mtcars, lambda = .5)

# Create a basic perspective plot
persp(mid, xvar = "wt", yvar = "hp")

# Customize the plot by passing arguments to graphics:::persp.default()
persp(mid, "wt", "hp", theta = 210, phi = 20, col = "lightblue", shade = .5)
persp(mid, "factor(am):wt", theta = 210, shade = .2)
}
\seealso{
\code{\link[graphics]{persp}}
}
