% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastLmMatrix.R
\name{fastLmMatrix}
\alias{fastLmMatrix}
\alias{fastLmMatrix.default}
\alias{fastLmMatrix.formula}
\alias{set_fastLmMatrix}
\title{Fit Multivariate Linear Models}
\usage{
fastLmMatrix(x, ...)

\method{fastLmMatrix}{default}(x, y, tol = 1e-07, method = 0L, ...)

\method{fastLmMatrix}{formula}(formula, data = list(), method = 0L, ...)

set_fastLmMatrix(which = 0L:4L)
}
\arguments{
\item{x}{a model matrix \eqn{X}.}

\item{...}{optional parameters passed to methods.}

\item{y}{the response matrix \eqn{Y}.}

\item{tol}{tolerance for the rank calculation.}

\item{method}{an integer with value \code{0} for the column-pivoted QR decomposition, \code{1} for the unpivoted QR decomposition, \code{2} for the LLT Cholesky, \code{3} for the LDLT Cholesky, and \code{4} for the Jacobi singular value decomposition (SVD). Default is zero.}

\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by \code{as.data.frame} to a data frame) containing the variables in the model.}

\item{which}{an integer vector specifying the methods to be registered. If the vector has names, they are used to construct the option names (e.g., \code{"midr.solver.(name)"}).}
}
\value{
\code{fastLmMatrix()} returns a list with the following components:
\item{coefficients}{\eqn{p \times k} matrix of coefficients.}
\item{fitted.values}{\eqn{n \times k} matrix of fitted values.}
\item{residuals}{\eqn{n \times k} matrix of residuals.}
\item{rank}{an integer giving the numeric rank of the model matrix \eqn{X}.}
}
\description{
\code{fastLmMatrix()} estimates the linear model for multivariate response using one of several methods implemented using the \code{Eigen} linear algebra library.
}
\details{
\code{fastLmMatrix()} is a performance-optimized version of the standard \code{lm} function, specifically designed to handle multivariate responses (\eqn{Y} as a matrix).
It leverages the \code{Eigen} C++ template library for high-performance linear algebra, utilizing \code{Eigen::Map} to avoid unnecessary memory copies and providing several decomposition methods with different trade-offs between speed and numerical stability.

\code{set_fastLmMatrix()} registers the C++ based internal functions as the OLS solvers for \code{midr::interpret()} via global \code{options}.
}
\examples{
# Multivariate Regression with Anscombe's Quartet
# Regress three 'y' variables on 'x1' simultaneously
Y <- as.matrix(anscombe[, c("y1", "y2", "y3")])
X <- as.matrix(cbind(1, anscombe$x1)) # Include intercept

fit_ans <- fastLmMatrix(X, Y, method = 0L)
print(fit_ans$coefficients)

# Formula interface with cbind()
fit_form <- fastLmMatrix(cbind(y1, y2, y3) ~ x1, data = anscombe)
print(fit_form$coefficients)
}
\keyword{internal}
