% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_reg.R
\name{mid_reg}
\alias{mid_reg}
\alias{fit_mid_reg}
\title{MID Regression}
\usage{
mid_reg(
  mode = "regression",
  engine = "midr",
  model = NULL,
  params_main = NULL,
  params_inter = NULL,
  penalty = NULL,
  terms = NULL
)

fit_mid_reg(
  formula,
  data,
  weights = NULL,
  model = NULL,
  params_main = NULL,
  params_inter = NULL,
  penalty = NULL,
  terms = NULL,
  ...
)
}
\arguments{
\item{mode}{a single character string for the type of model. Currently, only "regression" is supported.}

\item{engine}{a single character string specifying the computational engine to use. The default is "midr".}

\item{model}{an optional fitted model object (black-box model) to be interpreted. Default is \code{NULL}.}

\item{params_main}{an integer specifying the maximum number of sample points (knots) to model main effects. Corresponds to the argument \code{k[1]} in \code{midr::interpret()}.}

\item{params_inter}{an integer specifying the maximum number of sample points (knots) to model interaction effects. Corresponds to the argument \code{k[2]} or \code{k2} in \code{midr::interpret()}.}

\item{penalty}{a non-negative number representing the total amount of regularization. Corresponds to the argument \code{lambda} in \code{midr::interpret()}.}

\item{terms}{a character vector or formula of term labels to be included in the fitting process.}

\item{formula}{an object of class \code{formula}: a symbolic description of the model to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{weights}{an optional vector of case weights.}

\item{...}{other arguments to be passed on to \code{midr::interpret()}.}
}
\value{
a "model_spec" object with class "mid_reg".
}
\description{
\code{mid_reg()} defines a model that can predict numeric values from predictors using a set of functions each with up to two predictors.
This function fits a regression model based on the Maximum Interpretation Decomposition.
}
\details{
This function is the main specification for the \strong{parsnip} model.
The underlying fitting is performed by \code{fit_mid_reg()}, which is a helper function that connects \code{fit()} to the \code{midr::interpret()} function.
}
\seealso{
\code{\link[midr]{interpret}}
}
\keyword{internal}
