% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_surrogate.R
\name{mid_surrogate}
\alias{mid_surrogate}
\title{Global Surrogate Model Specification}
\usage{
mid_surrogate(
  mode = "regression",
  params_main = NULL,
  params_inter = NULL,
  penalty = NULL,
  custom_formula = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.}

\item{params_main}{An integer for the maximum number of sample points to model main effects (`k` for main effects in midr).}

\item{params_inter}{An integer for the maximum number of sample points to model interaction effects (`k` for interactions in midr).}

\item{penalty}{A non-negative number representing the total amount of regularization (`lambda` in midr).}
}
\description{
\code{mid_surrogate()} defines a model that serves as a surrogate model of another target model. This function can fit classification and regression models.
}
