% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midnight-package.R
\docType{package}
\name{midnight-package}
\alias{midnight}
\alias{midnight-package}
\title{midnight: A Tidymodels Interface for the 'midr' Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'midnight' package implements a 'parsnip' engine for the 'midr' package, allowing users to seamlessly fit, tune, and evaluate MID (Maximum Interpretation Decomposition) models with 'tidymodels' workflows. Development and augmentation of the package are driven by research from the "Moonlight Seminar 2025", a collaborative study group of actuaries from the Institute of Actuaries of Japan focused on advancing the practical applications of interpretable models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ryo-asashi/midnight}
  \item \url{https://ryo-asashi.github.io/midnight/}
  \item Report bugs at \url{https://github.com/ryo-asashi/midnight/issues}
}

}
\author{
\strong{Maintainer}: Ryoichi Asasihba \email{ryoichi.asashiba@gmail.com}

}
\keyword{internal}
